+++ src/chart/rendering/drawaxes.cpp
@@ -13,6 +13,7 @@
 #include "base/gfx/colortransform.h"
 #include "base/gfx/font.h"
 #include "base/math/fuzzybool.h"
+#include "base/math/renard.h"
 #include "base/type/booliter.h"
 #include "chart/generator/plot.h" // NOLINT(misc-include-cleaner)
 #include "chart/main/events.h"
@@ -21,6 +22,7 @@
 #include "dataframe/old/types.h"
 
 #include "drawguides.h"
+#include "drawinterlacing.h"
 #include "drawlabel.h"
 #include "orientedlabel.h"
 #include "renderedchart.h"
@@ -30,17 +32,20 @@ namespace Vizzu::Draw
 
 void DrawAxes::drawGeometries() const
 {
-	interlacing.drawGeometries();
+	DrawInterlacing{*this}.drawGeometries(Gen::AxisId::y);
+	DrawInterlacing{*this}.drawGeometries(Gen::AxisId::x);
 
 	drawAxis(Gen::AxisId::x);
 	drawAxis(Gen::AxisId::y);
 
-	DrawGuides{{ctx()}, canvas, painter}.draw();
+	DrawGuides{*this}.draw(Gen::AxisId::x);
+	DrawGuides{*this}.draw(Gen::AxisId::y);
 }
 
 void DrawAxes::drawLabels() const
 {
-	interlacing.drawTexts();
+	DrawInterlacing{*this}.drawTexts(Gen::AxisId::y);
+	DrawInterlacing{*this}.drawTexts(Gen::AxisId::x);
 
 	drawDimensionLabels(Gen::AxisId::x);
 	drawDimensionLabels(Gen::AxisId::y);
@@ -49,9 +54,165 @@ void DrawAxes::drawLabels() const
 	drawTitle(Gen::AxisId::y);
 }
 
-Geom::Line DrawAxes::getAxis(Gen::AxisId axisIndex) const
+const DrawAxes &&DrawAxes::init() &&
 {
-	auto offset = plot->axises.other(axisIndex).measure.origo();
+	for (auto axisIndex : Refl::enum_values<Gen::AxisId>()) {
+		const auto &axis = getAxis(axisIndex);
+
+		auto measEnabled = axis.measure.enabled.combine<double>();
+		auto &intervals = this->intervals[axisIndex];
+		auto &separators = this->separators[axisIndex];
+		const auto &guides = plot->guides.at(axisIndex);
+
+		for (auto &&[index, item] : axis.dimension.getValues()) {
+			auto weight = item.weight(axis.dimension.factor);
+			if (Math::Floating::is_zero(weight)) continue;
+
+			auto needInterlacing =
+			    measEnabled == 0.0
+			    || Math::FuzzyBool::And(Math::FuzzyBool::more(weight),
+			           guides.interlacings.more())
+			           != false;
+
+			intervals.emplace_back(item.range,
+			    weight,
+			    Math::FuzzyBool::And<double>(
+			        Math::Niebloid::interpolate(
+			            (item.startPos ? *item.startPos
+			                           : *item.endPos)
+			                % 2,
+			            (item.endPos ? *item.endPos : *item.startPos)
+			                % 2,
+			            axis.dimension.factor),
+			        needInterlacing),
+			    Interval::DimLabel{index,
+			        item.label,
+			        !item.startPos.isAuto(),
+			        !item.endPos.isAuto()});
+
+			auto needSeparators =
+			    measEnabled == 0.0
+			    || Math::FuzzyBool::And(Math::FuzzyBool::more(weight),
+			           Math::FuzzyBool::Or(guides.axisSticks.more(),
+			               guides.axisGuides.more()))
+			           != false;
+
+			if (auto sepWeight = Math::Niebloid::interpolate(
+			        !item.startPos.isAuto() && *item.startPos,
+			        !item.endPos.isAuto() && *item.endPos,
+			        axis.dimension.factor);
+			    needSeparators && sepWeight > 0)
+				separators.emplace_back(item.range.getMin(),
+				    sepWeight);
+		}
+
+		if (measEnabled == 0.0) continue;
+		auto step = axis.measure.step.combine();
+
+		auto &&[min, max] = std::minmax(
+		    axis.measure.step.get_or_first(::Anim::first).value,
+		    axis.measure.step.get_or_first(::Anim::second).value,
+		    Math::Floating::less);
+
+		auto stepHigh =
+		    std::clamp(Math::Renard::R5().ceil(step), min, max);
+		auto stepLow =
+		    std::clamp(Math::Renard::R5().floor(step), min, max);
+
+		if (Math::Floating::is_zero(axis.measure.range.size()))
+			step = stepHigh = stepLow = 1.0;
+
+		if (stepHigh == step || stepLow == step)
+			generateMeasure(axisIndex, step, measEnabled);
+		else {
+			auto highWeight =
+			    Math::Range<>::Raw(stepLow, stepHigh).rescale(step);
+
+			generateMeasure(axisIndex,
+			    stepLow,
+			    (1.0 - highWeight) * measEnabled);
+			generateMeasure(axisIndex,
+			    stepHigh,
+			    highWeight * measEnabled);
+		}
+	}
+
+	return std::move(*this);
+}
+
+void DrawAxes::generateMeasure(Gen::AxisId axisIndex,
+    double stepSize,
+    double weight)
+{
+	auto orientation = !+axisIndex;
+	const auto &meas = getAxis(axisIndex).measure;
+	auto rangeSize = meas.range.size();
+	auto singleLabelRange = Math::Floating::is_zero(rangeSize);
+
+	double stripWidth{};
+	if (!singleLabelRange) {
+		stripWidth = stepSize / rangeSize;
+		if (stripWidth <= 0) return;
+	}
+
+	const auto origo = this->origo();
+	auto axisBottom = origo.getCoord(!orientation) + stripWidth;
+
+	auto iMin = static_cast<int>(
+	    axisBottom > 0 ? std::floor(-origo.getCoord(!orientation)
+	                                / (2 * stripWidth))
+	                         * 2
+	                   : std::round((meas.range.getMin() - stepSize)
+	                                / stepSize));
+
+	if (axisBottom + iMin * stripWidth + stripWidth < 0.0)
+		iMin += 2
+		      * static_cast<int>(std::ceil(
+		          -(axisBottom + stripWidth) / (2 * stripWidth)));
+
+	auto Transform = [&](int x)
+	{
+		return std::pair{iMin + x * 2,
+		    axisBottom + (iMin + x * 2) * stripWidth};
+	};
+	constexpr static auto Predicate =
+	    [](const std::pair<int, double> &x)
+	{
+		return x.second <= 1.0;
+	};
+
+	auto &separators = this->separators[axisIndex];
+	auto &intervals = this->intervals[axisIndex];
+
+	for (auto &&[i, bottom] :
+	    std::views::iota(0) | std::views::transform(Transform)
+	        | std::views::take_while(Predicate)) {
+		if (!std::signbit(bottom)) {
+			separators.emplace_back(bottom,
+			    weight,
+			    (i + 1) * stepSize);
+		}
+
+		if (!singleLabelRange) {
+			intervals.emplace_back(
+			    Math::Range<>::Raw(bottom, bottom + stripWidth),
+			    weight,
+			    1.0);
+
+			if (!Math::Floating::less(1.0, bottom + stripWidth)) {
+				separators.emplace_back(bottom + stripWidth,
+				    weight,
+				    (i + 2) * stepSize);
+			}
+		}
+		else
+			break;
+	}
+}
+
+Geom::Line DrawAxes::getAxisLine(Gen::AxisId axisIndex) const
+{
+	auto offset = this->origo().getCoord(+!axisIndex);
 
 	auto direction = Geom::Point::Ident(+axisIndex);
 
@@ -64,7 +225,7 @@ Geom::Line DrawAxes::getAxis(Gen::AxisId axisIndex) const
 
 void DrawAxes::drawAxis(Gen::AxisId axisIndex) const
 {
-	if (auto line = getAxis(axisIndex); !line.isPoint()) {
+	if (auto line = getAxisLine(axisIndex); !line.isPoint()) {
 		auto lineColor =
 		    *rootStyle.plot.getAxis(axisIndex).color
 		    * static_cast<double>(plot->guides.at(axisIndex).axis);
@@ -104,9 +265,7 @@ Geom::Point DrawAxes::getTitleBasePos(Gen::AxisId axisIndex,
 	default:
 	case Pos::min_edge: break;
 	case Pos::max_edge: orthogonal = 1.0; break;
-	case Pos::axis:
-		orthogonal = plot->axises.other(axisIndex).measure.origo();
-		break;
+	case Pos::axis: orthogonal = origo().getCoord(+!axisIndex); break;
 	}
 
 	double parallel{0.0};
@@ -166,7 +325,7 @@ Geom::Point DrawAxes::getTitleOffset(Gen::AxisId axisIndex,
 
 void DrawAxes::drawTitle(Gen::AxisId axisIndex) const
 {
-	const auto &titleString = plot->axises.at(axisIndex).title;
+	const auto &titleString = getAxis(axisIndex).title;
 
 	const auto &titleStyle = rootStyle.plot.getAxis(axisIndex).title;
 
@@ -256,28 +415,27 @@ void DrawAxes::drawDimensionLabels(Gen::AxisId axisIndex) const
 	auto textColor = *labelStyle.color;
 	if (textColor.isTransparent()) return;
 
-	auto origo = plot->axises.origo();
-	const auto &axises = plot->axises;
-	const auto &axis = axises.at(axisIndex).dimension;
+	auto origo = this->origo();
+	const auto &axis = getAxis(axisIndex).dimension;
 
 	if (!axis.empty()) {
 		canvas.setFont(Gfx::Font{labelStyle});
 		const auto &enabled = plot->guides.at(axisIndex);
 
-		for (const auto &[slice, item] : axis.getValues())
+		for (const auto &interval : getIntervals(axisIndex)) {
+			if (!interval.label) continue;
 			drawDimensionLabel(axisIndex,
 			    origo,
-			    item,
-			    slice,
-			    Math::FuzzyBool::And<double>(item.weight(axis.factor),
+			    interval,
+			    Math::FuzzyBool::And<double>(interval.weight,
 			        enabled.labels));
+		}
 	}
 }
 
 void DrawAxes::drawDimensionLabel(Gen::AxisId axisIndex,
     const Geom::Point &origo,
-    const Gen::DimensionAxis::Item &item,
-    const Data::SliceIndex &index,
+    const Interval &interval,
     double weight) const
 {
 	if (weight == 0) return;
@@ -291,18 +449,17 @@ void DrawAxes::drawDimensionLabel(Gen::AxisId axisIndex,
 	        &axisIndex,
 	        &drawLabel,
 	        &labelStyle,
-	        &item,
+	        &interval,
 	        &orientation,
 	        &origo,
 	        ident = Geom::Point::Ident(orientation),
 	        normal = Geom::Point::Ident(!orientation),
-	        &text = item.label,
-	        &weight,
-	        &sindex = index](::Anim::InterpolateIndex index,
+	        &dimInfo = *interval.label,
+	        &weight](::Anim::InterpolateIndex index,
 	        const auto &position)
 	    {
 		    if (labelStyle.position->interpolates()
-		        && !item.presentAt(index))
+		        && !dimInfo.presentAt(index))
 			    return;
 
 		    Geom::Point refPos;
@@ -315,7 +472,7 @@ void DrawAxes::drawDimensionLabel(Gen::AxisId axisIndex,
 		    case Pos::min_edge: refPos = Geom::Point(); break;
 		    }
 
-		    auto relCenter = refPos + ident * item.range.middle();
+		    auto relCenter = refPos + ident * interval.range.middle();
 
 		    auto under =
 		        labelStyle.position->interpolates()
@@ -335,7 +492,7 @@ void DrawAxes::drawDimensionLabel(Gen::AxisId axisIndex,
 		    {
 			    if (!str.value) return;
 			    drawLabel.draw(canvas,
-			        sindex.value,
+			        dimInfo.index.value,
 			        posDir,
 			        labelStyle,
 			        0,
@@ -344,22 +501,22 @@ void DrawAxes::drawDimensionLabel(Gen::AxisId axisIndex,
 			                str.weight,
 			                plusWeight)),
 			        *rootEvents.draw.plot.axis.label,
-			        Events::Targets::dimAxisLabel(sindex.column,
-			            sindex.value,
+			        Events::Targets::dimAxisLabel(
+			            dimInfo.index.column,
+			            dimInfo.index.value,
 			            axisIndex));
 		    };
 
-		    if (text.interpolates()
+		    if (dimInfo.presented.interpolates()
 		        && !labelStyle.position->interpolates()) {
-			    draw(text.get_or_first(::Anim::first));
-			    draw(text.get_or_first(::Anim::second));
+			    draw(dimInfo.presented.get_or_first(::Anim::first));
+			    draw(dimInfo.presented.get_or_first(::Anim::second));
 		    }
-		    else {
-			    draw(text.get_or_first(index),
+		    else
+			    draw(dimInfo.presented.get_or_first(index),
 			        !labelStyle.position->interpolates()
 			            ? 1.0
 			            : position.weight);
-		    }
 	    });
 }